+++ server/app/views/user/user.scala.html
@@ -50,11 +50,11 @@
       <table class="table table-condensed">
         <thead>
           <tr>
-            <th>艦種</th><th>名前</th><th>Lv</th><th>Exp</th><th>HP</th><th>Cond</th><th>火力</th><th>雷装</th><th>対空</th><th>装甲</th><th>回避</th><th>対潜</th><th>索敵</th><th>運</th>
+            <th>艦種</th><th>名前</th><th>Lv</th><th>Exp</th><th>HP</th><th>Cond</th><th>火力</th><th>雷装</th><th>対空</th><th>装甲</th><th>回避</th><th>対潜</th><th>索敵</th><th>運</th><th>嫁</th>
           </tr>
         </thead>
         <tbody>
-        @for(ship <- yomes ++ best ++ flagship) {
+        @for((ship, isYome) <- yomes.map(_ -> true) ++ (best ++ flagship).map(_ -> false)) {
           <tr>
             <td>@ship.stName</td>
             <td>@ship.name</td>
@@ -70,6 +70,11 @@
             <td>@ship.taisen</td>
             <td>@ship.sakuteki</td>
             <td>@ship.lucky</td>
+            @if(isYome && user.isMine) {
+              <td>
+                <button type="submit" class="btn btn-warning btn-xs delete-yome" data-ship-id="@ship.id">嫁解除</button>
+              </td>
+            }
           </tr>
         }
         </tbody>
