+++ server/conf/db/migration/default/V61__add_unique_yome.sql
@@ -1,0 +1,2 @@
+
+alter table yome_ship add unique index (member_id, ship_id);
